Added reset score func to modal
